Fix OneDrive and SharePoint tests
