+++ src/Makefile
@@ -10,7 +10,6 @@ BAD_LINT_MSG := "Missing golangci-lint version $(WANTED_LINT_VERSION). Visit $(I
 
 build:
 	go build -o corso
-	echo Build complete
 
 lint: check-lint-version
 	golangci-lint run
