+++ src/internal/m365/collection/drive/helper_test.go
@@ -1,4 +1,4 @@
-package onedrive
+package drive
 
 import (
 	"testing"
