+++ src/internal/m365/service/onedrive/stub/stub.go
@@ -8,8 +8,8 @@ import (
 	"github.com/alcionai/clues"
 	"github.com/google/uuid"
 
+	"github.com/alcionai/corso/src/internal/m365/collection/drive/metadata"
 	odConsts "github.com/alcionai/corso/src/internal/m365/service/onedrive/consts"
-	"github.com/alcionai/corso/src/internal/m365/service/onedrive/metadata"
 	m365Stub "github.com/alcionai/corso/src/internal/m365/stub"
 	"github.com/alcionai/corso/src/internal/version"
 	"github.com/alcionai/corso/src/pkg/path"
