+++ src/internal/m365/service/sharepoint/helper_test.go
@@ -1,15 +1,9 @@
 package sharepoint
 
 import (
-	"testing"
-
-	"github.com/alcionai/clues"
 	msgraphsdk "github.com/microsoftgraph/msgraph-sdk-go"
-	"github.com/stretchr/testify/require"
 
-	"github.com/alcionai/corso/src/internal/m365/graph"
 	"github.com/alcionai/corso/src/internal/m365/support"
-	"github.com/alcionai/corso/src/pkg/account"
 )
 
 // ---------------------------------------------------------------------------
@@ -40,18 +34,4 @@ func (ms *MockGraphService) Adapter() *msgraphsdk.GraphRequestAdapter {
 }
 
 func (ms *MockGraphService) UpdateStatus(*support.ControllerOperationStatus) {
-}
-
-// ---------------------------------------------------------------------------
-// Helper Functions
-// ---------------------------------------------------------------------------
-
-func createTestService(t *testing.T, credentials account.M365Config) *graph.Service {
-	adapter, err := graph.CreateAdapter(
-		credentials.AzureTenantID,
-		credentials.AzureClientID,
-		credentials.AzureClientSecret)
-	require.NoError(t, err, "creating microsoft graph service for exchange", clues.ToCore(err))
-
-	return graph.NewService(adapter)
 }
