+++ src/internal/operations/test/onedrive_test.go
@@ -20,9 +20,9 @@ import (
 	"github.com/alcionai/corso/src/internal/events"
 	evmock "github.com/alcionai/corso/src/internal/events/mock"
 	"github.com/alcionai/corso/src/internal/m365"
+	"github.com/alcionai/corso/src/internal/m365/collection/drive"
+	"github.com/alcionai/corso/src/internal/m365/collection/drive/metadata"
 	"github.com/alcionai/corso/src/internal/m365/graph"
-	"github.com/alcionai/corso/src/internal/m365/service/onedrive"
-	"github.com/alcionai/corso/src/internal/m365/service/onedrive/metadata"
 	"github.com/alcionai/corso/src/internal/m365/resource"
 	"github.com/alcionai/corso/src/internal/model"
 	"github.com/alcionai/corso/src/internal/streamstore"
@@ -132,8 +132,8 @@ func (suite *OneDriveBackupIntgSuite) TestBackup_Run_incrementalOneDrive() {
 		return id
 	}
 
-	grh := func(ac api.Client) onedrive.RestoreHandler {
-		return onedrive.NewRestoreHandler(ac)
+	grh := func(ac api.Client) drive.RestoreHandler {
+		return drive.NewRestoreHandler(ac)
 	}
 
 	runDriveIncrementalTest(
@@ -157,7 +157,7 @@ func runDriveIncrementalTest(
 	category path.CategoryType,
 	includeContainers func([]string) selectors.Selector,
 	getTestDriveID func(*testing.T, context.Context) string,
-	getRestoreHandler func(api.Client) onedrive.RestoreHandler,
+	getRestoreHandler func(api.Client) drive.RestoreHandler,
 	skipPermissionsTests bool,
 ) {
 	t := suite.T()
@@ -388,7 +388,7 @@ func runDriveIncrementalTest(
 		{
 			name: "add permission to new file",
 			updateFiles: func(t *testing.T, ctx context.Context) {
-				err = onedrive.UpdatePermissions(
+				err = drive.UpdatePermissions(
 					ctx,
 					rh,
 					driveID,
@@ -406,7 +406,7 @@ func runDriveIncrementalTest(
 		{
 			name: "remove permission from new file",
 			updateFiles: func(t *testing.T, ctx context.Context) {
-				err = onedrive.UpdatePermissions(
+				err = drive.UpdatePermissions(
 					ctx,
 					rh,
 					driveID,
@@ -425,7 +425,7 @@ func runDriveIncrementalTest(
 			name: "add permission to container",
 			updateFiles: func(t *testing.T, ctx context.Context) {
 				targetContainer := containerInfos[container1].id
-				err = onedrive.UpdatePermissions(
+				err = drive.UpdatePermissions(
 					ctx,
 					rh,
 					driveID,
@@ -444,7 +444,7 @@ func runDriveIncrementalTest(
 			name: "remove permission from container",
 			updateFiles: func(t *testing.T, ctx context.Context) {
 				targetContainer := containerInfos[container1].id
-				err = onedrive.UpdatePermissions(
+				err = drive.UpdatePermissions(
 					ctx,
 					rh,
 					driveID,
